requirements.txt
assemblyai
httpx
numpy
openrouter>=0.1.1
psutil
pyaudio
python-dotenv
textblob
websockets
nltk
fastapi
uvicorn
pydantic
mcp
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client